requirements.txt
streamlit>=1.28.0
transformers>=4.30.0
torch>=2.0.0
anthropic>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0
python-docx>=1.0.0